Inserindo uma exceção no delete

+++ src/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { CreateUserDTO } from "./dto/create-user.dto";
 import { PrismaService } from "src/prisma/prisma.service";
 import { UpdatePutUserDTO } from "./dto/update-put-user.dto";
@@ -50,6 +50,11 @@ export class UserService {
     }
 
     async delete(id:number){
+
+        if (!(await this.show(id))){
+            throw new NotFoundException(`O usuário ${id} não existe.`);
+        }
+        
         return this.prisma.user.delete({
             where: {
                 id
